requirements.txt
asgiref==3.2.7
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
cryptography==3.4.3
dj-database-url==0.5.0
Django==3.1.7
django-channels==0.7.0
django-classy-tags==2.0.0
django-cms==3.8.0
django-cors-middleware==1.5.0
django-filer==2.0.2
django-filter==2.4.0
django-formtools==2.2
django-js-asset==1.2.2
django-markdown==0.8.4
django-mptt==0.11.0
django-oauth-toolkit==1.3.3
django-polymorphic==3.0.0
django-redis==4.12.1
django-rest-knox==4.1.0
django-sekizai==2.0.0
django-taggit==1.3.0
django-treebeard==4.4
django-widget-tweaks==1.4.8
djangocms-admin-style==2.0.2
djangocms-attributes-field==2.0.0
djangocms-bootstrap4==2.0.0
djangocms-file==3.0.0
djangocms-googlemap==2.0.0
djangocms-icon==2.0.0
djangocms-installer==2.0.0
djangocms-link==3.0.0
djangocms-picture==3.0.0
djangocms-style==3.0.0
djangocms-text-ckeditor==4.0.0
djangocms-video==3.0.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
easy-thumbnails==2.7.1
filter==0.0.0.20200724
html5lib==1.1
idna==2.10
Markdown==3.3.3
marshmallow==3.10.0
mysqlclient==2.0.3
oauthlib==3.1.0
Pillow==8.1.0
pycparser==2.20
PyJWT==2.0.1
pytz==2020.1
redis==3.5.3
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.3.1
tzlocal==2.1
Unidecode==1.1.2
urllib3==1.26.2
webencodings==0.5.1
